refac: alterado tipo para Long

+++ picpaysimplificado/src/main/java/com/picpaysimplificado/dtos/TransactionDTO.java
@@ -2,5 +2,5 @@ package com.picpaysimplificado.dtos;
 
 import java.math.BigDecimal;
 
-public record TransactionDTO(BigDecimal amount, Integer senderId, Integer receiverId) {
+public record TransactionDTO(BigDecimal amount, Long senderId, Long receiverId) {
 }
